Snapshot Management of VMWare with PowerShell
This is one of those management tasks that comes up at any location
you’re at, especially when you’re trying to manage VMs, performance or
datastore space. The cleanup task of deleting snapshots is easy, but the
questions that always comes to mind are: Who created it, when and why.
Take a look at a quick script I wrote to answer that information. I
schedule it to run every Monday morning and email the results. Simple!
This framework works for both VMware and Hyper-V.