Setting SQL Server Memory Allocation Maximum and Minimum with PowerShell
If you’ve ever run an installation of SQL Server, you’ll know it’s a
database, and databases love, love, love memory. By design and by
default, Microsoft SQL Server thinks it’s the only process on the system
and is therefore given all the available memory and CPUs. As a best
practice, I limit this. Here is the script I use to edit these values.