Using PowerShell to Add a Direct Member to an SCCM Collection
I was working at a client site and was going through their server
rollout procedure. I was quite shocked as to how many manual tasks they
still had. One of these tasks was to add a computer directly to a SCCM
collection. According to their requirements, they had to use direct
membership and could not do a WMI call. So, I created the following
script and added it to their task sequence.