Using PowerShell to Update the .Default and All User Profiles Registry
There are times that you may need to push out a change to all existing
user profiles and to new profiles that are created on a system. I’ve
seen a few PowerShell scripts floating around out there, but they didn’t
seem to work for Windows 7 SP1. You may or may not be surprised, but
there are many organizations that still run Windows 7. The script is
actually pretty simple.