What's the difference between -split and .split() in PowerShell?
Here’s a question I see over and over and over again: “I have a string
and I’m trying to split it on this part, but it’s jumbling it into a big
mess. What’s going on?” Well, there’s splitting a string in PowerShell,
and then there’s splitting a string in PowerShell. Confused? Let me
explain.