The Difference Between Get-Member and .GetType() in PowerShell
Recently, I was helping someone in a forum who was trying to figure out
what kind of object their command was returning. They knew about the
standard cmdlets people suggest when you’re getting started (Get-Help,
Get-Member, and Get-Command), but couldn’t figure out what was coming
back from a specific command.
In order to make this a more generic example, and to simplify it, let’s
approach this differently. Say I have these two objects where one is a
string and the other is an array of two strings.