PowerShell Rules for Format-Table and Format-List 
In PowerShell, when outputting data to the console, it’s typically
either organized into a table or a list. You can force output to take
either of these forms using the Format-Table and the Format-List
cmdlets, and people who write PowerShell cmdlets and modules can take
special steps to make sure their output is formatted as they desire.
But, when no developer has specifically asked for a formatted output,
how does PowerShell choose to display a table or a list?