Add a Work Note to a ServiceNow Incident with PowerShell
Recently, I’ve been working more with ServiceNow and writing scripts and
tools which sometimes interact with it. One of the things that I find
myself doing a lot is using PowerShell to add a work note to an
incident. Luckily, ServiceNow has an API that you can use to interact
with it and do this (among many other things).