Import Active Directory Module into Windows PE
One thing I have been wanting to have is access to active directory in a
WinPE environment. The main reason I want it is to be able to delete
systems from active directory during a build. When I first started
researching, I found this blog that guided me on writing this script.
The blog tells how to inject the AD module into the WIM file. That is
fine, but do you really want to do that every time you generate a new
WIM file? I don't. I started testing to see if the directories could be
copied into the WinPE environment while it was running without the need
of a reboot. It worked. Currently, this script only makes the Active
Directory module available in the WinPE environment. I am going to write
more scripts to take advantage of the AD module.