Set Windows Features and Verify with PowerShell
I am in the beginning stages of creating a Windows 10 build. One of the
first things I needed to do was to install and set the Windows 10
features. Before, I used a batch script that executed DISM to set each
feature. I know there is the Install-WindowsFeatures cmdlet, but I also
wanted to incorporate verification and other features into a single
script.