Uninstall an Application by Name with PowerShell
Here is a function that will uninstall an MSI installed application by
the name of the app. You do not need to input the entire name either.
For instance, say you are uninstalling all previous versions of Adobe
Reader. Adobe Reader is always labeled Adobe Reader X, Adobe Reader XI,
and so forth. This script allows you to do this without having to find
out every version that is installed throughout a network and then enter
an uninstaller line for each version. You just need to enter Adobe
Reader as the application name and the desired switches. It will then
search the name fields in the 32 and 64 bit uninstall registry keys to
find the associated GUID. Finally, it will execute an msiexec.exe /x
{GUID} to uninstall that version.