Utilizing Webhook Data in Functions and Validate Results using PowerShell
It's been a long road, but we're almost there!  A couple of weeks ago we
looked at how we can create an Azure Automation Account and add our own
custom modules to the solution to be used in Azure Automation.  Last
week, we took a deeper dive into configuring a runbook to take in
webhook data from an alert using Microsoft's Operations Management
Suite.  Then we looked into the data itself to see how we can leverage
it against our runbook to fix problems for us on the fly.