Onboarding Automation DSC Endpoints and Reporting
In the last chapter we talked about modifying and uploading our
configurations to Azure Automation DSC.  We were able to import
credentials from Azure's Automation Account Credential store, and then
compile the .mof files in the automation account for deployment.  In
this chapter, we'll be looking at how we apply those configurations to
existing systems via PowerShell.  Then we'll take a look at some of the
reporting available via Azure Automation DSC and send those reports over
to Operations Management Suite for dashboarding.