Publishing Configurations and Pushing them with Azure DSC
So, we've talked about Azure Automation DSC and the extensive reporting
we can get from it.  With the pricing as it is, it would be hard to
argue as to why you would want to use anything else.  But I'm a
completionist, and there may be some edge cases that might come up where
you wouldn't be able to use the pull method for configurations.  So
let's talk about how you can use Azure DSC to push a configuration to a
virtual machine.